fully implemented the form models an the relation between user and forms
Since the two to many relationship between users and forms doesn't natively work in JPA, I seperate the forms of a user to senderForms and receiverForms which turns the twotoMany relationship into 2 one to many relationships. Also had to turn the inheritence type of the form class to joined for this to work.

Also implemented the items of forms by creating an item class for forms which hold the equivalent class pairs and unidriectionally joined.